🐛 [fix] TextEditor UI 수정
